@@ -1,10 +1,48 @@
-// https://docs.expo.dev/guides/using-eslint/
-const { defineConfig } = require('eslint/config');
-const expoConfig = require('eslint-config-expo/flat');
+import typescriptEslintPlugin from '@typescript-eslint/eslint-plugin';
+import typescriptEslintParser from '@typescript-eslint/parser';
+import eslintPluginImport from 'eslint-plugin-import';
+import eslintPluginJsxA11y from 'eslint-plugin-jsx-a11y';
+import eslintPluginPrettier from 'eslint-plugin-prettier';
+import eslintPluginReact from 'eslint-plugin-react';
+import eslintPluginReactHooks from 'eslint-plugin-react-hooks';
 
-module.exports = defineConfig([
-  expoConfig,
+export default [
   {
-    ignores: ['dist/*'],
+    ignores: ['node_modules', 'build', 'dist'], // Ignorar diretórios
   },
-]);
+  {
+    files: ['**/*.ts', '**/*.tsx'], // Aplicar regras a arquivos TypeScript/React
+    languageOptions: {
+      parser: typescriptEslintParser,
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true,
+        },
+        ecmaVersion: 'latest',
+        sourceType: 'module',
+      },
+    },
+    plugins: {
+      react: eslintPluginReact,
+      'react-hooks': eslintPluginReactHooks,
+      import: eslintPluginImport,
+      'jsx-a11y': eslintPluginJsxA11y,
+      prettier: eslintPluginPrettier,
+      '@typescript-eslint': typescriptEslintPlugin,
+    },
+    rules: {
+      'prettier/prettier': 'error',
+      'react/react-in-jsx-scope': 'off',
+      'react/prop-types': 'off',
+      '@typescript-eslint/explicit-module-boundary-types': 'off',
+      '@typescript-eslint/no-unused-vars': [
+        'error',
+        { argsIgnorePattern: '^_' },
+      ],
+      'import/prefer-default-export': 'off',
+      'react/jsx-props-no-spreading': 'off',
+      'no-use-before-define': 'off',
+      '@typescript-eslint/no-use-before-define': ['off'],
+    },
+  },
+];
